# language-school-app/src/app/api/model/commentDTO.ts
/**
 * Language School API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CommentDTO { 
    text?: string;
    userId?: number;
    taskId?: number;
    privateStatus?: boolean;
    id?: number;
}