# language-school-app/src/app/api/model/groupAnswerDTO.ts
/**
 * Language School API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LanguageDTO } from './languageDTO';


export interface GroupAnswerDTO { 
    name: string;
    description?: string;
    difficulty?: GroupAnswerDTO.DifficultyEnum;
    language?: LanguageDTO;
    id?: number;
}
export namespace GroupAnswerDTO {
    export type DifficultyEnum = 'BEGINNER' | 'ELEMENTARY' | 'UPPER_INTERMEDIATE' | 'ADVANCED' | 'PROFICIENCY';
    export const DifficultyEnum = {
        Beginner: 'BEGINNER' as DifficultyEnum,
        Elementary: 'ELEMENTARY' as DifficultyEnum,
        UpperIntermediate: 'UPPER_INTERMEDIATE' as DifficultyEnum,
        Advanced: 'ADVANCED' as DifficultyEnum,
        Proficiency: 'PROFICIENCY' as DifficultyEnum
    };
}