# language-school-app/src/openapi.json
{"openapi":"3.1.0","info":{"title":"Language School API","version":"1.0"},"servers":[{"url":"http://worker.thallassianangel.su","description":"Generated server url"}],"security":[{"JWT":[]}],"paths":{"/task/{taskId}/edit":{"put":{"tags":["task-controller"],"operationId":"editTask","parameters":[{"name":"taskId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaskDTO"}}}}}}},"/language/{languageId}/edit":{"put":{"tags":["language-controller"],"operationId":"editLanguageName","parameters":[{"name":"languageId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LanguageDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LanguageDTO"}}}}}}},"/group/{groupId}/edit":{"put":{"tags":["group-controller"],"operationId":"editGroup","parameters":[{"name":"groupId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GroupAnswerDTO"}}}}}}},"/comment/{commentId}/edit":{"put":{"tags":["comment-controller"],"operationId":"editComment","parameters":[{"name":"commentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}}}}}},"/api/users/teachers/{id}":{"get":{"tags":["user-controller"],"operationId":"getTeacherById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserFullDTO"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateTeacher","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeacherUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteTeacher","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/api/users/students/{id}":{"get":{"tags":["user-controller"],"operationId":"getStudentById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserFullDTO"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateStudent","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteStudent","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/task/{taskId}/complete_task":{"post":{"tags":["task-controller"],"operationId":"completeTask","parameters":[{"name":"taskId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/task/create":{"post":{"tags":["task-controller"],"operationId":"createTask","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaskDTO"}}}}}}},"/notification/create":{"post":{"tags":["notification-controller"],"operationId":"createNotification","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationCreationModel"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotificationDto"}}}}}}},"/language/create":{"post":{"tags":["language-controller"],"operationId":"createLanguage","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LanguageDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LanguageDTO"}}}}}}},"/group/{groupId}/add/{userId}":{"post":{"tags":["group-controller"],"operationId":"addStudentToGroup","parameters":[{"name":"groupId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GroupAnswerDTO"}}}}}},"delete":{"tags":["group-controller"],"operationId":"removeStudentFromGroup","parameters":[{"name":"groupId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GroupAnswerDTO"}}}}}}},"/group/create":{"post":{"tags":["group-controller"],"operationId":"createGroup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GroupAnswerDTO"}}}}}}},"/comment/create":{"post":{"tags":["comment-controller"],"operationId":"createComment","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}}}}}},"/auth/register/teacher":{"post":{"tags":["auth-controller"],"operationId":"registerTeacher","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}}}}},"/auth/register/student":{"post":{"tags":["auth-controller"],"operationId":"registerStudent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterStudentDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}}}}},"/auth/register/admin":{"post":{"tags":["auth-controller"],"operationId":"registerAdmin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}}}}},"/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}}}}},"/attachments":{"post":{"tags":["attachment-controller"],"operationId":"uploadAttachment","parameters":[{"name":"taskId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK"}}}},"/attachments/to-notification":{"post":{"tags":["attachment-controller"],"operationId":"uploadAttachmentForNotification","parameters":[{"name":"taskId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK"}}}},"/api/users/teachers":{"get":{"tags":["user-controller"],"operationId":"getAllTeachers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"post":{"tags":["user-controller"],"operationId":"createTeacher","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeacherCreateDTO"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/api/users/students":{"get":{"tags":["user-controller"],"operationId":"getAllStudents","parameters":[{"name":"groupId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"post":{"tags":["user-controller"],"operationId":"createStudent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentCreateDTO"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/task/{teacherId}/get_by_teacher":{"get":{"tags":["task-controller"],"operationId":"getByTeacherId","parameters":[{"name":"teacherId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskTeacherDTO"}}}}}}}},"/task/get_by_group_name_real":{"get":{"tags":["task-controller"],"operationId":"getByGroupNameNew","parameters":[{"name":"groupName","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskStudentDTO"}}}}}}}},"/task/get_by_group_name":{"get":{"tags":["task-controller"],"operationId":"getByGroupName","parameters":[{"name":"groupName","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskStudentDTO"}}}}}}}},"/notification/for-students/{studentId}":{"get":{"tags":["notification-controller"],"operationId":"getAllStudentsNotifications","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NotificationDto"}}}}}}}},"/notification/by-group/{groupId}":{"get":{"tags":["notification-controller"],"operationId":"getAllGroupNotifications","parameters":[{"name":"groupId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NotificationDto"}}}}}}}},"/language/get_all_languages":{"get":{"tags":["language-controller"],"operationId":"getLanguages","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LanguageDTO"}}}}}}}},"/group/{teacherId}/get_groups_by_teacher":{"get":{"tags":["group-controller"],"operationId":"getGroupsByTeacherId","parameters":[{"name":"teacherId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GroupAnswerDTO"}}}}}}}},"/group/{groupId}/get_group_by_id":{"get":{"tags":["group-controller"],"operationId":"getByGroupId","parameters":[{"name":"groupId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GroupAnswerDTO"}}}}}}},"/group/get_all_groups":{"get":{"tags":["group-controller"],"operationId":"getGroups","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GroupAnswerDTO"}}}}}}}},"/group/get-groups-with-filters":{"get":{"tags":["group-controller"],"operationId":"getWithFilters","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"difficulty","in":"query","required":false,"schema":{"type":"string","enum":["BEGINNER","ELEMENTARY","UPPER_INTERMEDIATE","ADVANCED","PROFICIENCY"]}},{"name":"language","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GroupAnswerDTO"}}}}}}}},"/comment/{taskId}/get":{"get":{"tags":["comment-controller"],"operationId":"getCommentsByTaskId","parameters":[{"name":"taskId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CommentDTO"}}}}}}}},"/attachments/{attachmentId}/download":{"get":{"tags":["attachment-controller"],"operationId":"downloadAttachment","parameters":[{"name":"attachmentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}}}}},"/api/users":{"get":{"tags":["user-controller"],"operationId":"getUsers","parameters":[{"name":"groupId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}}}}},"/api/users/students/search":{"get":{"tags":["user-controller"],"operationId":"searchStudents","parameters":[{"name":"value","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}}}}},"/api/users/me":{"get":{"tags":["user-controller"],"operationId":"getMe","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserFullDTO"}}}}}}},"/task/{taskId}/delete":{"delete":{"tags":["task-controller"],"operationId":"deleteTask","parameters":[{"name":"taskId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/language/{languageId}/delete":{"delete":{"tags":["language-controller"],"operationId":"deleteLanguage","parameters":[{"name":"languageId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/group/{groupId}/delete":{"delete":{"tags":["group-controller"],"operationId":"deleteGroup","parameters":[{"name":"groupId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/comment/{commentId}/delete":{"delete":{"tags":["comment-controller"],"operationId":"deleteComment","parameters":[{"name":"commentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/attachments/{attachmentId}":{"delete":{"tags":["attachment-controller"],"operationId":"deleteAttachment","parameters":[{"name":"attachmentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"TaskDTO":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"deadline":{"type":"string","format":"date"},"groupName":{"type":"string"}}},"LanguageDTO":{"type":"object","properties":{"name":{"type":"string"}}},"GroupDTO":{"type":"object","properties":{"name":{"type":"string","minLength":1},"description":{"type":"string"},"difficulty":{"type":"string","enum":["BEGINNER","ELEMENTARY","UPPER_INTERMEDIATE","ADVANCED","PROFICIENCY"]},"language":{"$ref":"#/components/schemas/LanguageDTO"}},"required":["name"]},"GroupAnswerDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string","minLength":1},"description":{"type":"string"},"difficulty":{"type":"string","enum":["BEGINNER","ELEMENTARY","UPPER_INTERMEDIATE","ADVANCED","PROFICIENCY"]},"language":{"$ref":"#/components/schemas/LanguageDTO"}},"required":["name"]},"CommentDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"text":{"type":"string"},"userId":{"type":"integer","format":"int64"},"taskId":{"type":"integer","format":"int64"},"privateStatus":{"type":"boolean"}}},"TeacherUpdateDTO":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"}}},"UserDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"groups":{"type":"array","items":{"$ref":"#/components/schemas/GroupAnswerDTO"}},"role":{"type":"string","enum":["TEACHER","STUDENT","ADMIN"]}}},"StudentUpdateDTO":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"grade":{"type":"string"}}},"NotificationCreationModel":{"type":"object","properties":{"text":{"type":"string"},"groupId":{"type":"integer","format":"int64"},"files":{"type":"array","items":{"type":"string","format":"binary"}}}},"AttachmentDownloadInfo":{"type":"object","properties":{"fileName":{"type":"string"},"fileType":{"type":"string"},"fileSize":{"type":"integer","format":"int64"},"objectKey":{"type":"string"}}},"NotificationDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"text":{"type":"string"},"groupId":{"type":"integer","format":"int64"},"creationDate":{"type":"string","format":"date"},"createdByTeacherWithId":{"type":"integer","format":"int64"},"attachments":{"type":"array","items":{"$ref":"#/components/schemas/AttachmentDownloadInfo"}}}},"RegisterDTO":{"type":"object","properties":{"firstName":{"type":"string","minLength":1},"lastName":{"type":"string","minLength":1},"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","maxLength":64,"minLength":8}},"required":["email","firstName","lastName","password"]},"TokenDTO":{"type":"object","properties":{"token":{"type":"string","minLength":1}},"required":["token"]},"RegisterStudentDTO":{"type":"object","properties":{"firstName":{"type":"string","minLength":1},"lastName":{"type":"string","minLength":1},"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","maxLength":64,"minLength":8},"grade":{"type":"string"}},"required":["email","firstName","lastName","password"]},"LoginDTO":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"TeacherCreateDTO":{"type":"object","properties":{"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"password":{"type":"string"}}},"StudentCreateDTO":{"type":"object","properties":{"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"password":{"type":"string"},"grade":{"type":"string"},"groupIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"TaskTeacherDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"deadline":{"type":"string","format":"date"},"commentList":{"type":"array","items":{"$ref":"#/components/schemas/CommentDTO"}},"attachmentDownloadInfos":{"type":"array","items":{"$ref":"#/components/schemas/AttachmentDownloadInfo"}}}},"TaskStudentDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"deadline":{"type":"string","format":"date"},"taskStatus":{"type":"string","enum":["COMPLETE","OVERDUE","PENDING"]},"teacher":{"$ref":"#/components/schemas/UserDTO"}}},"UserFullDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"groups":{"type":"array","items":{"$ref":"#/components/schemas/GroupAnswerDTO"}},"role":{"type":"string","enum":["TEACHER","STUDENT","ADMIN"]}}}},"securitySchemes":{"JWT":{"type":"http","name":"JWT","scheme":"bearer","bearerFormat":"JWT"}}}}